CLN: extract Implicit function
